auto_report/requirements.txt
pandas
schedule
reportlab
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
matplotlib
python-dateutil